+++ CHANGELOG.md
@@ -14,6 +14,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 ## [2.3.2] - 12/04/21
 ### Added
 - @JvmStatic annotation in CropImage.activity() and fun activity(uri) [#108](https://github.com/CanHub/Android-Image-Cropper/issues/108)
+- @JvmStatic annotation in all static function in  CropImage [#108](https://github.com/CanHub/Android-Image-Cropper/issues/108)
 
 ## [2.3.1] - 01/04/21
 ### Changed